fix: Unit tests added, code refactoring
