+++ src/components/Avatar/avatar.component.tsx
