+++ src/components/Button/test/Button.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Button from '../';
+import Button from '../../Button';
 import BR from '@blueeast/bluerain-os';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
