+++ src/components/Card/index.tsx
@@ -1,5 +1,5 @@
-export {default as  CardHeader } from './card/CardHeader.component';
-export {default as  CardMedia } from './card/CardMedia.component';
-export {default as  CardBody } from './card/CardBody.component';
-export {default as  CardFooter } from './card/CardFooter.component';
-export { default } from './card/Card.component';
+export { default as  CardHeader } from './CardHeader.component';
+export { default as  CardMedia } from './CardMedia.component';
+export { default as  CardBody } from './CardBody.component';
+export { default as  CardFooter } from './CardFooter.component';
+export { default } from './Card.component';
