+++ src/components/Chip/README.md
