+++ src/components/Chip/index.ts
@@ -1,4 +1,4 @@
 /**
  * Created by Ghazal Taimur on 12/29/17.
  */
-export { default } from './chip';
+export { default } from './Chip.component';
