+++ src/components/DropDown/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './DropDown';
+export { default } from './DropDown.component';
