+++ src/components/Icon/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Icon';
+export { default } from './Icon.component';
