+++ src/components/List/index.tsx
@@ -1,8 +1,8 @@
-import ListItem from './list/ListItem.component';
-import ListItemAvatar from './list/ListItemAvatar.component';
-import ListItemText from './list/ListItemText.component';
-import ListItemRightButton from './list/ListItemRightButton.component';
-import ListItemIcon from './list/ListItemIcon.component';
-import ListHeader from './list/ListHeader.component';
-export { default } from './list/List.component';
+import ListItem from './ListItem.component';
+import ListItemAvatar from './ListItemAvatar.component';
+import ListItemText from './ListItemText.component';
+import ListItemRightButton from './ListItemRightButton.component';
+import ListItemIcon from './ListItemIcon.component';
+import ListHeader from './ListHeader.component';
+export { default } from './List.component';
 export { ListItem, ListItemAvatar, ListItemIcon, ListHeader, ListItemText, ListItemRightButton };
