+++ src/components/Switch/Switch.component.tsx
