+++ src/components/Switch/test/Switch.stories.tsx
@@ -1,22 +1,22 @@
 import * as React from 'react';
-import Switch from '../../switch';
+import Switch from '../../Switch';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { withKnobs, boolean, number} from '@storybook/addon-knobs';
+import { withKnobs, boolean, number } from '@storybook/addon-knobs';
 const stories = storiesOf('Switch', module);
 stories.add('Switch (with knobs)', () => {
-    const disabled = boolean('disabled', false);
-    const value = boolean('value', true);
-    return <Switch disabled={disabled} value={value} onValueChange={action('Value is',value)}/>;
+	const disabled = boolean('disabled', false);
+	const value = boolean('value', true);
+	return <Switch disabled={disabled} value={value} onValueChange={action('Value is',value)}/>;
 });
 stories.add('Switch (default)', () => {
-    return  (
+	return  (
     <Switch />);
 });
 
 stories.add('Switch (MUI specific props)', () => {
-    return  (
-<div style={{display:'flex',}}>
+	return  (
+<div style={{ display:'flex', }}>
 <Switch aria-label="checkedD" />
         <Switch
           checked={true}
