--- src/components/badge/index.ts
@@ -1,1 +1,0 @@
-export { default } from './badge';
